+++ src/payloads/edit_message_caption.rs
@@ -15,7 +15,7 @@ impl_payload! {
             /// Unique identifier for the target chat or username of the target channel (in the format `@channelusername`).
             pub chat_id: ChatId [into],
             /// Identifier of the message to edit
-            pub message_id: i64,
+            pub message_id: i32,
             /// New caption of the message, 0-1024 characters after entities parsing
             pub caption: String [into],
         }
