+++ src/payloads/stop_message_live_location.rs
@@ -18,7 +18,7 @@ impl_payload! {
             /// Unique identifier for the target chat or username of the target channel (in the format `@channelusername`)
             pub chat_id: ChatId [into],
             /// Identifier of the message to edit
-            pub message_id: i64,
+            pub message_id: i32,
             /// Latitude of new location
             pub latitude: f64,
             /// Longitude of new location
